refactor(export): address second-pass review feedback
- Sanitize failureReason in the UI: persist a generic user-facing message
  and rely on the logger for the full exception details (no more raw
  Doctrine / filesystem error text leaking into the profile page).
- Fix file_size column/PHP type mismatch: column is now INTEGER (signed
  32-bit is ample for any plausible ZIP) so MySQL no longer returns a
  string into the `?int` property.
- Validate context keys/values in ExportAction with the same regex used
  for gridFilters; reject non-scalar values up front instead of trusting
  whatever JSON the client sent.
- Defense-in-depth assertion in CompanyExporter — fail loudly if no
  active company is set so non-CompanyAware children can't silently
  leak across tenants via findAll().
- Replace `@filesize()` with an explicit is_file + filesize so genuine
  I/O failures surface instead of being swallowed.
- Swap predictable uniqid() for bin2hex(random_bytes(8)) in the staging
  directory name.
- GridQueryService now skips SortFilter construction entirely on empty
  sort, with a comment explaining the SortFilter empty-field tolerance.
- Document the deliberate decoupling of ExportJob.requestedBy from the
  UserBundle (bare ULID + DB-level FK rather than an ORM relation).
- Extract a shared ValueFormatter so GridRowExtractor and
  EntityRowNormalizer agree on Money/DateTime/Enum/Ulid/scalar
  handling. Each caller still owns its BigNumber/null branches.
- Drop the emoji from the export_ready email per CLAUDE.md.
- Add ExportJobVoter unit tests covering grant, two denial branches,
  no-active-company, and the two abstain cases.

+++ migrations/Version30000_8.php
@@ -43,7 +43,7 @@ final class Version30000_8 extends AbstractMigration
         $table->addColumn('format', Types::STRING, ['length' => 10, 'notnull' => true]);
         $table->addColumn('status', Types::STRING, ['length' => 20, 'notnull' => true]);
         $table->addColumn('archive_path', Types::STRING, ['length' => 512, 'notnull' => false]);
-        $table->addColumn('file_size', Types::BIGINT, ['notnull' => false]);
+        $table->addColumn('file_size', Types::INTEGER, ['notnull' => false]);
         $table->addColumn('created_at', Types::DATETIME_IMMUTABLE, ['notnull' => true]);
         $table->addColumn('completed_at', Types::DATETIME_IMMUTABLE, ['notnull' => false]);
         $table->addColumn('failure_reason', Types::TEXT, ['notnull' => false]);
